requirements.txt
pyyaml
gensim
pandas
numpy
datasets
nltk
tqdm
scikit-learn
matplotlib
